pyproject.toml
[project]
name = "kebele-agent"
version = "0.1.0"
description = "🇪🇹 AI Agent for Kebele Services (Hackathon MVP)"
requires-python = ">=3.12"
dependencies = [
  "gradio==4.26.0",
  "langchain==0.1.0",
  "langchain-openai==0.0.5",
  "langgraph==0.0.14",
  "python-dotenv==1.0.0",
  "openai>=1.0.0",
  "pydantic==2.5.0",
  "huggingface-hub<1.0",
  "typer>=0.9.0",
  "reportlab",
]
